# pixiforma-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pixiforma",
        "description": "Documentation de l'API REST Pixiforma",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/formations/{formation}/block-requests": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "Liste des demandes de déblocage (formateur)",
                "operationId": "df7d55573c8a5d4a88af1a241badac88",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des demandes"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Soumettre une demande de déblocage",
                "operationId": "1ebe10d345651abec4e663750efa89f7",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Demande envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Non bloqué ou demande déjà existante"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/block-requests/{blockRequest}/pass": {
            "put": {
                "tags": [
                    "Formations"
                ],
                "summary": "Valider la publication pour l'apprenant (formateur)",
                "operationId": "2499436e8f038ba2f22ebc9d5e875d4f",
                "parameters": [
                    {
                        "name": "blockRequest",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Publication validée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/formations/{formation}/publications": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Associer des publications à une formation",
                "operationId": "1b8975e521cafe407ef51422c2ac7da1",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublicationIdsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Publications associées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Liste des notifications",
                "description": "Retourne la liste des notifications de l'utilisateur connecté avec le nombre de notifications non lues.",
                "operationId": "0f031f66cd8c52bf73e89fb2d050ad83",
                "responses": {
                    "200": {
                        "description": "Liste des notifications récupérée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/verify-reset-token": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier la validité d'un token de réinitialisation",
                "description": "Vérifie si un token de réinitialisation de mot de passe est valide et non expiré.",
                "operationId": "7819899b7198f2d4bee4e3d0cba942dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@test.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token valide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Token invalide ou expiré"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Réinitialisation du mot de passe",
                "operationId": "62ed51acd0b89e4f5411eb8e9d8c1462",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "Password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe modifié avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lien invalide ou données incorrectes"
                    }
                }
            }
        },
...
        "/users/{user}/role": {
            "put": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Attribuer un rôle à un utilisateur",
                "operationId": "ee25d1a97fb55365530e85db33e35d4b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôle attribué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Opération réussie."
                    },
                    "data": {
                        "type": "mixed",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PublicationIdsRequest": {
                "properties": {
                    "publication_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "durations": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "BlockRequest": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "formation_user_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "formation_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "declined"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "nullable": true
                    },
                    "formateur_response": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AssignRoleRequest": {
                "properties": {
                    "roles": {
                        "description": "Rôles à assigner",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "Admin",
                                "Formateur",
                                "Collaborateur"
                            ]
                        },
                        "example": [
                            "Collaborateur"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Jeton d'authentification Bearer. Format : Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}